Add files via upload
This is the application which uses technology .NET Core framework. It has architecture MVC developed by the pattern. It uses the Northwind database, and it consists of "Products" and "Category" Entity class.  In constructor HomeControllers the application calls injected context from Northwind ContextDb class, then in ActionResults put data in lists. In the major page in the View Index.cshtml  I used combination C#, HTML  and CSS